Authorization - Creating a Policy

+++ app/Http/Controllers/TaskController.php
@@ -69,6 +69,10 @@ class TaskController extends Controller
      */
     public function destroy(Request $request, Task $task)
     {
+        $this->authorize('destroy', $task);
+
         $task->delete();
+
+        return redirect('/tasks');
     }
 }
